Fix configuration setup
This commit fixes some issues with configuration setup in the plugin.
Firstly, it adds the service available option which was missing from
the plugin. (this opt will be removed from tempest when this lands in
openstack) This also fixes the variable name used for registering the
plugins which was previously incorrectly using BotoGroup.